feat: migrate major pages from client-side API calls to server-side data fetching
This commit implements a comprehensive migration of three major pages from
client-side API fetching to server-side data loading using Next.js App Router
best practices and the shared data layer.

## Pages Migrated

### My Submissions Page (/app/(creator)/my-submissions/)
- Convert from client-side fetch("/api/submissions") to server-side getCreatorSubmissions()
- Create my-submissions-client.tsx for interactive filtering and search
- Add proper authentication and user ID handling with getCurrentUser()
- Improve error handling and loading states

### Discover Page (/app/(creator)/discover/)
- Convert from client-side API calls to server-side getCampaigns()
- Create discover-client.tsx for interactive filtering, search, and bookmarks
- Enhance getCampaigns() to support featured campaigns and advanced filtering
- Maintain all interactive functionality while improving initial load performance

### IP Kits Page (/app/(brand)/ip-kits/)
- Convert from client-side fetch("/api/ip-kits") to server-side getIpKits()
- Create ip-kits-client.tsx for search and filter interactions
- Add proper brand permission checking with getUserBrandIds()
- Improve error handling for users without brand access

## Infrastructure Improvements

### Enhanced Shared Data Layer
- Add getCurrentUser() helper for server components authentication
- Enhance getCampaigns() with featured campaigns support (featured=true param)
- Add proper filtering for isNotNull() and advanced search capabilities
- Fix TypeScript type handling for null/undefined values in database responses

### Architecture Benefits
- **Performance**: Eliminate client-side data fetching waterfalls
- **SEO**: Server-side rendering with data improves initial page load
- **UX**: Faster perceived performance, reduced loading spinners
- **Architecture**: Clean separation of concerns following Next.js 13+ best practices
- **Consistency**: All major pages now follow the same data fetching pattern

## Quality Assurance
- TypeScript compilation passes without errors
- ESLint passes (warnings are for pre-existing image optimization issues)
- Maintained all existing functionality and interactive features
- Proper error boundaries and fallback states implemented
