+++ ai/implementation-plan.md
@@ -233,6 +233,23 @@ You will follow this exact sequence from the Implementation Plan:
 - [x] Ensure consistent filter state across page refreshes and navigation
 - [x] Test browser navigation scenarios thoroughly
 
+#### 4.8 Server-Side Data Fetching Migration (Feature #7c)
+
+**Completed: December 2024**
+
+- [x] Migrate my-submissions page from client-side fetch to server component using getCreatorSubmissions()
+- [x] Convert discover page from client-side API calls to server-side getCampaigns() with initial data loading
+- [x] Transform IP kits page from client-side fetch to server component using getIpKits()
+- [x] Create proper server/client component separation maintaining interactive functionality
+- [x] Add getCurrentUser() helper function for server-side authentication in components
+- [x] Enhance getCampaigns() shared data layer function with featured campaigns and advanced filtering
+- [x] Fix TypeScript type handling for null/undefined database values in transformations
+- [x] Implement proper error boundaries and fallback states for server components
+- [x] Add comprehensive authentication and brand permission checking in server components
+- [x] Eliminate unnecessary HTTP requests during server-side rendering for better performance
+- [x] Maintain all interactive features (search, filters, bookmarks) through client components
+- [x] Ensure TypeScript compilation and ESLint compliance across all converted components
+
 ### Phase 5: Canvas Implementation
 
 #### 5.1 Basic Canvas Infrastructure (Feature #8)
@@ -521,10 +538,14 @@ You will follow this exact sequence from the Implementation Plan:
 
 #### 11.1 Client Component Data Layer Migration
 
-- [ ] Convert my-submissions page to use getCreatorSubmissions() for better performance
-- [ ] Optimize discover page client-side data fetching with shared functions
-- [ ] Convert IP kits page to use getIpKits() shared function
-- [ ] Replace assets page API calls with shared data layer functions
+- [x] Convert my-submissions page to use getCreatorSubmissions() for better performance
+- [x] Optimize discover page client-side data fetching with shared functions
+- [x] Convert IP kits page to use getIpKits() shared function
+- [x] Replace assets page API calls with shared data layer functions
+- [x] Implement proper server/client component separation with Next.js App Router best practices
+- [x] Add getCurrentUser() helper function for server-side authentication
+- [x] Enhance shared data layer with featured campaigns and advanced filtering support
+- [x] Fix TypeScript type handling for null/undefined values in database responses
 - [ ] Implement client-side caching and optimistic updates
 - [ ] Add React Query or SWR for better client-side data management
 - [ ] Optimize re-rendering and data synchronization across components
